Arreglados el path, falta las tuberias

+++ Libft/sources/calloc.c
@@ -6,16 +6,16 @@
 /*   By: mvelazqu <mvelazqu@student.42barcel>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/03/28 17:04:03 by mvelazqu          #+#    #+#             */
-/*   Updated: 2024/03/28 17:06:34 by mvelazqu         ###   ########.fr       */
+/*   Updated: 2024/03/30 14:49:44 by mvelazqu         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include <stdlib.h>
 
-void	*ft_calloc(size_t count, size_t size)
+void	*ft_calloc(int count, int size)
 {
 	char	*rtrn;
-	size_t	i;
+	int		i;
 
 	rtrn = malloc(sizeof(char) * count * size);
 	if (!rtrn)
